refactor: #151 - useKeyword 사용자 훅 정의 및 비즈니스 로직 분리
