+++ src/components/SearchBar/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useKeyword } from './useKeyword';
